+++ tests/rest/conftest.py
@@ -8,6 +8,7 @@ from unittest import mock
 import pytest
 import time_machine
 
+from openeo.rest._testing import DummyBackend
 from openeo.rest.connection import Connection
 
 API_URL = "https://oeo.test/"
@@ -72,7 +73,19 @@ def oidc_device_code_flow_checker(time_machine, simple_time, fast_sleep, capsys)
 
 
 @pytest.fixture
+def con100(requests_mock):
+    requests_mock.get(API_URL, json={"api_version": "1.0.0"})
+    con = Connection(API_URL)
+    return con
+
+
+@pytest.fixture
 def con120(requests_mock):
     requests_mock.get(API_URL, json={"api_version": "1.2.0"})
     con = Connection(API_URL)
     return con
+
+
+@pytest.fixture
+def dummy_backend(requests_mock, con100) -> DummyBackend:
+    yield DummyBackend(requests_mock=requests_mock, connection=con100)