+++ tests/rest/datacube/test_datacube100.py
@@ -6,10 +6,11 @@ Unit tests specifically for 1.0.0-style DataCube
 import collections
 import copy
 import io
+import json
 import pathlib
+import sys
 import re
 import textwrap
-import warnings
 from typing import Optional
 
 import pytest
@@ -58,6 +59,134 @@ basic_geometry_types = [
             "geometry": {"type": "Polygon", "coordinates": (((1, 0), (1, 1), (0, 1), (0, 0), (1, 0)),)},
         },
     ),
+]
+
+
+WKT2_FOR_EPSG23631 = """
+PROJCRS["WGS 84 / UTM zone 31N",
+    BASEGEOGCRS["WGS 84",
+        ENSEMBLE["World Geodetic System 1984 ensemble",
+            MEMBER["World Geodetic System 1984 (Transit)"],
+            MEMBER["World Geodetic System 1984 (G730)"],
+            MEMBER["World Geodetic System 1984 (G873)"],
+            MEMBER["World Geodetic System 1984 (G1150)"],
+            MEMBER["World Geodetic System 1984 (G1674)"],
+            MEMBER["World Geodetic System 1984 (G1762)"],
+            MEMBER["World Geodetic System 1984 (G2139)"],
+            ELLIPSOID["WGS 84",6378137,298.257223563,
+                LENGTHUNIT["metre",1]],
+            ENSEMBLEACCURACY[2.0]],
+        PRIMEM["Greenwich",0,
+            ANGLEUNIT["degree",0.0174532925199433]],
+        ID["EPSG",4326]],
+    CONVERSION["UTM zone 31N",
+        METHOD["Transverse Mercator",
+            ID["EPSG",9807]],
+        PARAMETER["Latitude of natural origin",0,
+            ANGLEUNIT["degree",0.0174532925199433],
+            ID["EPSG",8801]],
+        PARAMETER["Longitude of natural origin",3,
+            ANGLEUNIT["degree",0.0174532925199433],
+            ID["EPSG",8802]],
+        PARAMETER["Scale factor at natural origin",0.9996,
+            SCALEUNIT["unity",1],
+            ID["EPSG",8805]],
+        PARAMETER["False easting",500000,
+            LENGTHUNIT["metre",1],
+            ID["EPSG",8806]],
+        PARAMETER["False northing",0,
+            LENGTHUNIT["metre",1],
+            ID["EPSG",8807]]],
+    CS[Cartesian,2],
+        AXIS["(E)",east,
+            ORDER[1],
+            LENGTHUNIT["metre",1]],
+        AXIS["(N)",north,
+            ORDER[2],
+            LENGTHUNIT["metre",1]],
+    USAGE[
+        SCOPE["Engineering survey, topographic mapping."],
+        AREA["Between 0°E and 6°E, northern hemisphere between equator and 84°N, onshore and offshore. Algeria. Andorra. Belgium. Benin. Burkina Faso. Denmark - North Sea. France. Germany - North Sea. Ghana. Luxembourg. Mali. Netherlands. Niger. Nigeria. Norway. Spain. Togo. United Kingdom (UK) - North Sea."],
+        BBOX[0,0,84,6]],
+    ID["EPSG",32631]]
+"""
+
+
+PROJJSON_FOR_EPSG23631 = {
+    "$schema": "https://proj.org/schemas/v0.4/projjson.schema.json",
+    "type": "ProjectedCRS",
+    "name": "WGS 84 / UTM zone 31N",
+    "base_crs": {
+        "name": "WGS 84",
+        "datum_ensemble": {
+            "name": "World Geodetic System 1984 ensemble",
+            "members": [
+                {"name": "World Geodetic System 1984 (Transit)", "id": {"authority": "EPSG", "code": 1166}},
+                {"name": "World Geodetic System 1984 (G730)", "id": {"authority": "EPSG", "code": 1152}},
+                {"name": "World Geodetic System 1984 (G873)", "id": {"authority": "EPSG", "code": 1153}},
+                {"name": "World Geodetic System 1984 (G1150)", "id": {"authority": "EPSG", "code": 1154}},
+                {"name": "World Geodetic System 1984 (G1674)", "id": {"authority": "EPSG", "code": 1155}},
+                {"name": "World Geodetic System 1984 (G1762)", "id": {"authority": "EPSG", "code": 1156}},
+                {"name": "World Geodetic System 1984 (G2139)", "id": {"authority": "EPSG", "code": 1309}},
+            ],
+            "ellipsoid": {"name": "WGS 84", "semi_major_axis": 6378137, "inverse_flattening": 298.257223563},
+            "accuracy": "2.0",
+            "id": {"authority": "EPSG", "code": 6326},
+        },
+        "coordinate_system": {
+            "subtype": "ellipsoidal",
+            "axis": [
+                {"name": "Geodetic latitude", "abbreviation": "Lat", "direction": "north", "unit": "degree"},
+                {"name": "Geodetic longitude", "abbreviation": "Lon", "direction": "east", "unit": "degree"},
+            ],
+        },
+        "id": {"authority": "EPSG", "code": 4326},
+    },
+    "conversion": {
+        "name": "UTM zone 31N",
+        "method": {"name": "Transverse Mercator", "id": {"authority": "EPSG", "code": 9807}},
+        "parameters": [
+            {
+                "name": "Latitude of natural origin",
+                "value": 0,
+                "unit": "degree",
+                "id": {"authority": "EPSG", "code": 8801},
+            },
+            {
+                "name": "Longitude of natural origin",
+                "value": 3,
+                "unit": "degree",
+                "id": {"authority": "EPSG", "code": 8802},
+            },
+            {
+                "name": "Scale factor at natural origin",
+                "value": 0.9996,
+                "unit": "unity",
+                "id": {"authority": "EPSG", "code": 8805},
+            },
+            {"name": "False easting", "value": 500000, "unit": "metre", "id": {"authority": "EPSG", "code": 8806}},
+            {"name": "False northing", "value": 0, "unit": "metre", "id": {"authority": "EPSG", "code": 8807}},
+        ],
+    },
+    "coordinate_system": {
+        "subtype": "Cartesian",
+        "axis": [
+            {"name": "Easting", "abbreviation": "E", "direction": "east", "unit": "metre"},
+            {"name": "Northing", "abbreviation": "N", "direction": "north", "unit": "metre"},
+        ],
+    },
+    "scope": "Engineering survey, topographic mapping.",
+    "area": "Between 0°E and 6°E, northern hemisphere between equator and 84°N, onshore and offshore. Algeria. Andorra. Belgium. Benin. Burkina Faso. Denmark - North Sea. France. Germany - North Sea. Ghana. Luxembourg. Mali. Netherlands. Niger. Nigeria. Norway. Spain. Togo. United Kingdom (UK) - North Sea.",
+    "bbox": {"south_latitude": 0, "west_longitude": 0, "north_latitude": 84, "east_longitude": 6},
+    "id": {"authority": "EPSG", "code": 32631},
+}
+
+
+CRS_VALUES_SUPPORTED_FOR_ALL_PYTHON_VERSIONS = [
+    "EPSG:32631",
+    "32631",
+    32631,
+    "+proj=utm +zone=31 +datum=WGS84 +units=m +no_defs",  # is also EPSG:32631, in proj format
 ]
 
 
@@ -276,12 +405,13 @@ def test_aggregate_spatial_types(con100: Connection, polygon, expected_geometrie
     }
 
 
-def test_aggregate_spatial_with_crs(con100: Connection, recwarn):
+@pytest.mark.parametrize("crs", CRS_VALUES_SUPPORTED_FOR_ALL_PYTHON_VERSIONS)
+def test_aggregate_spatial_with_crs(con100: Connection, recwarn, crs: str):
     img = con100.load_collection("S2")
     polygon = shapely.geometry.box(0, 0, 1, 1)
-    masked = img.aggregate_spatial(geometries=polygon, reducer="mean", crs="EPSG:32631")
+    masked = img.aggregate_spatial(geometries=polygon, reducer="mean", crs=crs)
     warnings = [str(w.message) for w in recwarn]
-    assert "Geometry with non-Lon-Lat CRS 'EPSG:32631' is only supported by specific back-ends." in warnings
+    assert f"Geometry with non-Lon-Lat CRS {crs!r} is only supported by specific back-ends." in warnings
     assert sorted(masked.flat_graph().keys()) == ["aggregatespatial1", "loadcollection1"]
     assert masked.flat_graph()["aggregatespatial1"] == {
         "process_id": "aggregate_spatial",
@@ -298,6 +428,86 @@ def test_aggregate_spatial_with_crs(con100: Connection, recwarn):
         },
         "result": True
     }
+
+
+@pytest.mark.skipif(sys.version_info < (3, 7), reason="WKT2 format not supported by pyproj 3.0 / python 3.6")
+def test_aggregate_spatial_with_crs_as_wkt(con100: Connection, recwarn):
+    """Separate test coverage for WKT, so we can skip it for Python3.6"""
+    crs = WKT2_FOR_EPSG23631
+    img = con100.load_collection("S2")
+    polygon = shapely.geometry.box(0, 0, 1, 1)
+    masked = img.aggregate_spatial(geometries=polygon, reducer="mean", crs=crs)
+    warnings = [str(w.message) for w in recwarn]
+    assert f"Geometry with non-Lon-Lat CRS {crs!r} is only supported by specific back-ends." in warnings
+    assert sorted(masked.flat_graph().keys()) == ["aggregatespatial1", "loadcollection1"]
+    assert masked.flat_graph()["aggregatespatial1"] == {
+        "process_id": "aggregate_spatial",
+        "arguments": {
+            "data": {"from_node": "loadcollection1"},
+            "geometries": {
+                "type": "Polygon",
+                "coordinates": (((1.0, 0.0), (1.0, 1.0), (0.0, 1.0), (0.0, 0.0), (1.0, 0.0)),),
+                "crs": {"properties": {"name": "EPSG:32631"}, "type": "name"},
+            },
+            "reducer": {
+                "process_graph": {
+                    "mean1": {"process_id": "mean", "arguments": {"data": {"from_parameter": "data"}}, "result": True}
+                }
+            },
+        },
+        "result": True,
+    }
+
+
+@pytest.mark.skipif(sys.version_info < (3, 8), reason="PROJJSON format not supported by pyproj 3.2 / python < v3.8")
+@pytest.mark.parametrize("crs", [PROJJSON_FOR_EPSG23631, json.dumps(PROJJSON_FOR_EPSG23631)])
+def test_aggregate_spatial_with_crs_as_projjson(con100: Connection, recwarn, crs):
+    """Separate test coverage for PROJJSON, so we can skip it for Python versions below 3.8"""
+    img = con100.load_collection("S2")
+    polygon = shapely.geometry.box(0, 0, 1, 1)
+    masked = img.aggregate_spatial(geometries=polygon, reducer="mean", crs=crs)
+    warnings = [str(w.message) for w in recwarn]
+    assert f"Geometry with non-Lon-Lat CRS {crs!r} is only supported by specific back-ends." in warnings
+    assert sorted(masked.flat_graph().keys()) == ["aggregatespatial1", "loadcollection1"]
+    assert masked.flat_graph()["aggregatespatial1"] == {
+        "process_id": "aggregate_spatial",
+        "arguments": {
+            "data": {"from_node": "loadcollection1"},
+            "geometries": {
+                "type": "Polygon",
+                "coordinates": (((1.0, 0.0), (1.0, 1.0), (0.0, 1.0), (0.0, 0.0), (1.0, 0.0)),),
+                "crs": {"properties": {"name": "EPSG:32631"}, "type": "name"},
+            },
+            "reducer": {
+                "process_graph": {
+                    "mean1": {"process_id": "mean", "arguments": {"data": {"from_parameter": "data"}}, "result": True}
+                }
+            },
+        },
+        "result": True,
+    }
+
+
+@pytest.mark.parametrize(
+    "crs",
+    [
+        "does-not-exist-crs",
+        "EEPSG:32165",  # Simulate a user typo
+        -1,  # negative value can not be valid EPSG code
+        "-1",  # string representing negative value can not be valid EPSG code
+        1.0,  # floating point: type is not supported
+        {1: 1},  # type is not supported
+        [1],  # type is not supported
+    ],
+)
+def test_aggregate_spatial_with_invalid_crs(con100: Connection, recwarn, crs: str):
+    """Test that it refuses invalid input for the CRS: incorrect types and negative integers,
+    i.e. things that can not be a CRS at all, soo it is not just a CRS proj does not know about.
+    """
+    img = con100.load_collection("S2")
+    polygon = shapely.geometry.box(0, 0, 1, 1)
+    with pytest.raises((ValueError, TypeError)):
+        img.aggregate_spatial(geometries=polygon, reducer="mean", crs=crs)
 
 
 def test_aggregate_spatial_target_dimension(con100: Connection):
@@ -476,12 +686,38 @@ def test_mask_polygon_types(con100: Connection, polygon, expected_mask):
     }
 
 
-def test_mask_polygon_with_crs(con100: Connection, recwarn):
+@pytest.mark.parametrize("crs", CRS_VALUES_SUPPORTED_FOR_ALL_PYTHON_VERSIONS)
+def test_mask_polygon_with_crs(con100: Connection, recwarn, crs: str):
     img = con100.load_collection("S2")
     polygon = shapely.geometry.box(0, 0, 1, 1)
-    masked = img.mask_polygon(mask=polygon, srs="EPSG:32631")
+    masked = img.mask_polygon(mask=polygon, srs=crs)
     warnings = [str(w.message) for w in recwarn]
-    assert "Geometry with non-Lon-Lat CRS 'EPSG:32631' is only supported by specific back-ends." in warnings
+    assert f"Geometry with non-Lon-Lat CRS {crs!r} is only supported by specific back-ends." in warnings
+    assert sorted(masked.flat_graph().keys()) == ["loadcollection1", "maskpolygon1"]
+    assert masked.flat_graph()["maskpolygon1"] == {
+        "process_id": "mask_polygon",
+        "arguments": {
+            "data": {"from_node": "loadcollection1"},
+            "mask": {
+                "type": "Polygon",
+                "coordinates": (((1.0, 0.0), (1.0, 1.0), (0.0, 1.0), (0.0, 0.0), (1.0, 0.0)),),
+                # All listed test inputs for crs should be converted to "EPSG:32631"
+                "crs": {"type": "name", "properties": {"name": "EPSG:32631"}},
+            },
+        },
+        "result": True,
+    }
+
+
+@pytest.mark.skipif(sys.version_info < (3, 7), reason="WKT2 format not supported by pyproj 3.0 / python 3.6")
+def test_mask_polygon_with_crs_as_wkt(con100: Connection, recwarn):
+    """Separate test coverage for WKT, so we can skip it for Python3.6"""
+    crs = WKT2_FOR_EPSG23631
+    img = con100.load_collection("S2")
+    polygon = shapely.geometry.box(0, 0, 1, 1)
+    masked = img.mask_polygon(mask=polygon, srs=crs)
+    warnings = [str(w.message) for w in recwarn]
+    assert f"Geometry with non-Lon-Lat CRS {crs!r} is only supported by specific back-ends." in warnings
     assert sorted(masked.flat_graph().keys()) == ["loadcollection1", "maskpolygon1"]
     assert masked.flat_graph()["maskpolygon1"] == {
         "process_id": "mask_polygon",
@@ -489,6 +725,32 @@ def test_mask_polygon_with_crs(con100: Connection, recwarn):
             "data": {"from_node": "loadcollection1"},
             "mask": {
                 "type": "Polygon", "coordinates": (((1.0, 0.0), (1.0, 1.0), (0.0, 1.0), (0.0, 0.0), (1.0, 0.0)),),
+                # All listed test inputs for crs should be converted to "EPSG:32631"
+                "crs": {"type": "name", "properties": {"name": "EPSG:32631"}},
+            },
+        },
+        "result": True,
+    }
+
+
+@pytest.mark.skipif(sys.version_info < (3, 8), reason="PROJJSON format not supported by pyproj 3.2 / python < v3.8")
+@pytest.mark.parametrize("crs", [PROJJSON_FOR_EPSG23631, json.dumps(PROJJSON_FOR_EPSG23631)])
+def test_mask_polygon_with_crs_as_projjson(con100: Connection, recwarn, crs):
+    """Separate test coverage for PROJJSON, so we can skip it for Python versions below 3.8"""
+    img = con100.load_collection("S2")
+    polygon = shapely.geometry.box(0, 0, 1, 1)
+    masked = img.mask_polygon(mask=polygon, srs=crs)
+    warnings = [str(w.message) for w in recwarn]
+    assert f"Geometry with non-Lon-Lat CRS {crs!r} is only supported by specific back-ends." in warnings
+    assert sorted(masked.flat_graph().keys()) == ["loadcollection1", "maskpolygon1"]
+    assert masked.flat_graph()["maskpolygon1"] == {
+        "process_id": "mask_polygon",
+        "arguments": {
+            "data": {"from_node": "loadcollection1"},
+            "mask": {
+                "type": "Polygon",
+                "coordinates": (((1.0, 0.0), (1.0, 1.0), (0.0, 1.0), (0.0, 0.0), (1.0, 0.0)),),
+                # All listed test inputs for crs should be converted to "EPSG:32631"
                 "crs": {"type": "name", "properties": {"name": "EPSG:32631"}},
             },
         },